examples/msal-python/requirements.txt
azure-keyvault-secrets==4.3.0
msal==1.14.0
